requirements.txt
flask
flask_sqlalchemy
sqlalchemy
pyodbc
azure-functions
azure-functions-wsgi
